change storage

+++ commands/_say.js
@@ -1,6 +1,7 @@
 const BotCommand = require('./BotCommand');
 const Logger = require('../terminal/Logger')
-const {fallback} = require("../handler/utils");
+const {fallback} = require('../handler/utils');
+const {owner} = require('../configurations/bot-settings.json');
 
 module.exports = new BotCommand(
     "say",
@@ -9,10 +10,11 @@ module.exports = new BotCommand(
     "say this is example",
     undefined,
     async (commandLabel, commandArgs, message, member, channel) => {
-        await message.delete().catch(Logger.error);
+        if (channel.type === "text")
+            await message.delete().catch(Logger.error);
         if (isNaN(commandArgs[0])) {
             await channel.send(commandArgs.join(" "));
-        } else {
+        } else if (member.id === owner || member.id === process["internal"]["discord"]["client"].id) {
             let cid = commandArgs.shift();
             fallback(
                 process["internal"]["discord"]["client"].channels.cache.get(cid),
